Adding groups in Escardó's style
There is a new file in Structures where groups are added as Martín Hötzel Escardó did in his Introduction to HoTT/UF with Agda (adding an extra axiom to Monoids). There is also a new lemma in Monoids for the uniqueness of inverse (when it exists).
Future plans: reduce the code, organize it and add groups with an inverse operation in the structure (and not defined from the axiom).