Newstructures (#284)
* Adding groups in Escardó's style

There is a new file in Structures where groups are added as Martín Hötzel Escardó did in his Introduction to HoTT/UF with Agda (adding an extra axiom to Monoids). There is also a new lemma in Monoids for the uniqueness of inverse (when it exists).
Future plans: reduce the code, organize it and add groups with an inverse operation in the structure (and not defined from the axiom).

* Add files via upload

Created semigroups, abelian groups and rings. Modified groups in order to make easier their equivalence. Lesson learned: if and operation can be defined naturally from an axiom, define it from there. Also added some useful functions to other files.

* Symmetric Groups

Starting a theory of symmetric groups

* Add files via upload

* Add files via upload

* Delete InftyMagma.agda

* Ring renamed to the more adequate name Rng

* Fixing trailing whitespace

* Fixing trailing whitespace

* Added everything file to Algebra directory

* Added Algebra.Everything to Readme.agda

* Trying to fix whitespace

* Changing local definition for function from the library

Changed the proof of a lemma for the use of the more general isOfHLevel\~- as Mortberg suggested.

* Preparing for the merge #1

Some changes suggested by Anders Mortberg:
- Changing plural to singular in Monoids, Groups, AbGroups to Monoid, Group, AbGroup
- Moved some computation rules of equivalences from Cubical.Algebra.SymmetricGroup to Cubical.Foundations.Equiv
- Eliminated raw-abelian-group-structure for raw-group-structure (they were defined as the same)
- Eliminated some definitions that are going to be added by another PR (Powersets)

* Preparing for merge #1

Some changes suggested by Anders Mortberg:
- Changing plural to singular in Monoids, Groups, AbGroups to Monoid, Group, AbGroup
- Moved some computation rules of equivalences from Cubical.Algebra.SymmetricGroup to Cubical.Foundations.Equiv
- Eliminated raw-abelian-group-structure for raw-group-structure (they were defined as the same)
- Eliminated some definitions that are going to be added by another PR (Powersets)

* Added modified Equiv.agda

I forgot to add the new Equiv.agda with more rules of how composition of equivalences compute (like composing inverses gives identity).

* Preparing for merge #2

Eliminate notation _\circ_ for compEquiv and moved \circ-assoc to Cubical.Foundations.Equiv renaming it as compEquiv-assoc

* Preparing for merge #2

Eliminate notation _\circ_ for compEquiv and moved \circ-assoc to Cubical.Foundations.Equiv renaming it as compEquiv-assoc

* Preparing for merge #2

Added explanation of the group axioms.

* Preparing for merge #3

Adding first version of Ring and CommRing, removing Rng.

* Preparing for merge #4

Added some new extractors and notations

* Preparing for merge #4

Added new extractors and syntax